+++ webworker/src/lib.rs
@@ -1,3 +1,4 @@
+#![recursion_limit = "256"]
 #![allow(clippy::too_many_arguments)]
 #![allow(clippy::cognitive_complexity)]
 
@@ -117,7 +118,7 @@ async fn handle(
       let state = state_option.as_mut().ok_or(anyhow!("Not initialized"))?;
       let confidence = InConfidence {
         minimax_depth: (8.0 * complexity).round() as u32,
-        uct_iterations: (100_000.0 * complexity).round() as usize,
+        uct_iterations: (100_000.0 * complexity).round() as u32,
         zero_iterations: (1_000.0 * complexity).round() as usize,
       };
       let analysis = state
